add: search component
