+++ src/components/common/pagination.vue
@@ -54,7 +54,7 @@ import { computed, defineComponent, onMounted, ref, toRefs, watch } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 
 export default defineComponent({
-  name: 'MyPagination',
+  name: 'SitePagination',
   props: {
     totalPages: {
       type: Number,
